requirements.txt
black==24.10.0
click==8.1.8
Faker==33.3.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pyperf==2.8.1
python-dateutil==2.9.0.post0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.12.2
